Image talons (#1803)
